refactor(context): invert DiagnosticComputer to drop the context -> diagnostics edge
DiagnosticComputer becomes an interface in context.computer
(compute(DocumentContext) -> List<Diagnostic>), and its implementation moves to
diagnostics.DefaultDiagnosticComputer, which iterates the BSLDiagnostic beans.
DocumentContext keeps injecting the interface by type, so Spring supplies the
implementation at runtime (DI wiring is not a compile-time package dependency).

This inverts the dependency: context now declares the contract and diagnostics
implements it (diagnostics -> context, the natural direction), so context no
longer imports the diagnostics engine. Combined with the OScriptModuleTypeResolver
move, context leaves the core cycle entirely: the per-package SCC collapses from
20 to 9 (residual core is types <-> references). The only remaining
context -> diagnostics reference is to the dependency-free metadata vocabulary.
